Update README.md
Clean Bin Agencies is here to protect health, well-being and the environment by providing reasonable measures for — minimizing the consumption of natural resources; avoiding and minimizing the generation of waste; reducing, re-using, recycling and recovering waste; treating and safely disposing of waste to ensure that peopleware aware of the impact of waste on their health, wellbeing and the environment. We want to teach the community on how to protect our environment and reduces the chances of spending deceases. Project outcome The community will gain knowledge and skills in proper disposal of hazardous waste, including sharp and other infectious waste that might affect health and environment. The community will have a safer environment and less prone to diseases. Project outputs1.Active cooperation of the